Übung OpenMP

+++ caches/matrixadd-soln.c
@@ -11,7 +11,7 @@ void **malloc2d(int x, int y, size_t size)
     void **array = malloc(x * sizeof(void *) + x * y * size);
     for (int i = 0; i < x; i++)
     {
-        int offset = x * sizeof(void *) + i * y * size;
+        size_t offset = x * sizeof(void *) + i * y * size;
         array[i] = (void *)((unsigned char *)array + offset);
     }
     return array;
